Moving rotate and translate and adding functions using them (fkin and inv_kin) to arm class
