+++ src/beachbot/manipulators/roarmm1.py
@@ -7,29 +7,6 @@ import json
 from scipy import signal
 import serial
 
-def rotate(point, angle, origin=(0, 0)):
-    """
-    Rotate a point counterclockwise by a given angle around a given origin.
-
-    The angle must be given in degrees.
-    """
-    ox, oy = origin
-    px, py = point
-
-    angle = angle * math.pi / 180
-
-    qx = ox + math.cos(angle) * (px - ox) - math.sin(angle) * (py - oy)
-    qy = oy + math.sin(angle) * (px - ox) + math.cos(angle) * (py - oy)
-    return qx, qy
-
-
-def translate(point, offset):
-    """
-    Translate point by offset.
-
-    """
-    return point[0] + offset[0], point[1] + offset[1]
-
 
 class RoArmM1(threading.Thread):
     def __init__(self, rate_hz=20, serial_port="/dev/ttyUSB0", gripper_limits=None) -> None:
@@ -309,68 +286,6 @@ class RoArmM1(threading.Thread):
     def go_home(self):
         self.set_joint_targets(self.q_home)
         time.sleep(1)
-
-    def fkin(self, qs):
-        """
-        x,y,z = fkin([angle_1, angle_2, angle_3, angle_4])
-        Calculate gripper position and rotation based on given joint angles
-
-        Code reconstructed from given inverse kinamtic routines provided by waveshare.
-        """
-        angle_1, angle_2, angle_3, angle_4 = qs
-
-        # First, calculate position in x/z plane (joint 2 - joint 4)
-        # Forward kinematic is calculated from tip of the robot backward to base!
-        # Start position:
-        p = (0, 0)
-        # gGripper offset translation:
-        p = translate(p, (self.LEN_E, -self.LEN_F))
-        # Gripper rotation (joint 4):
-        p = rotate(p, -angle_4)
-        # Translation from joint 4 to joint 3:
-        p = translate(p, (self.LEN_D, 0))
-        # Joint 3 rotation:
-        p = rotate(p, -angle_3)
-        # Translation from joint 3 to joint 2:
-        p = translate(p, (self.LEN_C, 0))
-        # Joint 2 Rotation
-        p = rotate(p, 90 - angle_2)
-        # Offset translation from joint 2 to base rotation
-        p = translate(p, (self.LEN_B, self.LEN_A))
-
-        # Rotate robot arm in x/y plane (joint 1):
-        p_plane = rotate((p[0], self.LEN_H), angle_1 - 180)
-
-        # Combine X/Z and X/Y plane caculation for final result:
-        return (p_plane[0], p_plane[1], p[1])
-
-    def inv_kin(self, position, eoat):
-        """
-        angle_1, angle_2, angle_3, angle_4 = inv_kin([InputX, InputY, InputZ], eoat)
-        Calculate required joint angles to reach a certain position/posture.
-        position: cartesian position
-        EOAT: desired End of Arm Tooling, rotation angle of gripper
-
-        Code provided by waveshare, rewritten in python
-        """
-        InputX, InputY, InputZ = position
-        InputTheta = eoat
-
-        # Rotation of the gripper (relative to arm) is defined as a constant fraction of the desired rotation:
-        # (ambiguity resolution)
-        InputTheta = eoat / self.rateTZ
-        # Joint 1 rotates the robot arm in the X/Y plane, calculate angles based on desired target X/Y position:
-        angle_1, len_totalXY = self.wigglePlaneIK(self.LEN_H, InputX, InputY)
-        # Calculate offset in XZ plane from last joint (joint 4) to gripper
-        angle_EoAT, len_a, len_b = self.EoAT_IK(InputTheta)
-        # Calculate required angles of arm joints (2,3) such that the target with desired rotation of the gripper cna be reached:
-        angle_2, angle_3, angle_IKE = self.simpleLinkageIK(
-            self.LEN_C, self.LEN_D, (len_totalXY - len_a), (InputZ - self.LEN_A + len_b)
-        )
-        # Estimate required joint angle of the gripper rotation (joint 4):
-        angle_4 = angle_IKE + angle_EoAT
-
-        return angle_1, angle_2, angle_3, angle_4
 
     def is_ready(self):
         """Robot arm connected and ready for operation"""
